Move more types from categories out
